refactor: move interfaces to bases and fix queue_status message format
- Rename interfaces/ directory to bases/ for better semantics
- Rename IInputAudioSession/IOutputAudioSession to BaseInputAudioSession/BaseOutputAudioSession
- Update all imports to use bases/ instead of interfaces/
- Fix queue_status message format to nest position/limit/active under data field
- Add TODO comment for audio enhancer in WebInputAudioSession
- Add action-handler.ts for handling conversation actions
